[FLINK-1201] [gelly] Renamed tests from TestXyz to XyzITCase
